[AN] refactor: Setting 화면 컴포저블 분리 #721
*   [AN] refactor: SettingScreen의 TopBar, Content 컴포저블 분리
*   [AN] refactor: 불필요한 openUrl 함수 제거
