+++ android/app/src/main/java/com/onair/hearit/presentation/setting/component/SettingContent.kt
@@ -1,7 +1,5 @@
-package com.onair.hearit.presentation.setting
+package com.onair.hearit.presentation.setting.component
 
-import android.content.Context
-import android.content.Intent
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
@@ -9,110 +7,32 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.livedata.observeAsState
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.core.net.toUri
-import com.google.android.gms.oss.licenses.OssLicensesMenuActivity
 import com.onair.hearit.R
 import com.onair.hearit.domain.model.UserInfo
 import com.onair.hearit.presentation.theme.Gray4
 import com.onair.hearit.presentation.theme.HearitBlack
-import com.onair.hearit.presentation.theme.HearitTypoGraphy
 import com.onair.hearit.presentation.theme.Red
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun SettingsScreen(
-    onBackClick: () -> Unit,
-    viewModel: SettingViewModel,
-) {
-    val context = LocalContext.current
-    val privacyPolicyUrl = stringResource(id = R.string.privacy_policy_url)
-    val termsUrl = stringResource(id = R.string.terms_of_use_url)
-    val userInfo by viewModel.userInfo.observeAsState(UserInfo.default())
-
-    Scaffold(
-        topBar = {
-            SettingsTopBar(onBackClick = onBackClick)
-        },
-        containerColor = HearitBlack,
-    ) { padding ->
-        SettingsContent(
-            modifier = Modifier.padding(padding),
-            userInfo = userInfo,
-            onProfileClick = { /* 내정보 화면으로 이동 */ },
-            onPrivacyPolicyClick = { openUrl(privacyPolicyUrl, context) },
-            onTermsClick = { openUrl(termsUrl, context) },
-            onOpenSourceClick = { navigateToLicense(context) },
-            onLogoutClick = { },
-            onWithdrawalClick = { },
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun SettingsTopBar(onBackClick: () -> Unit) {
-    TopAppBar(
-        title = {
-            Box(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(end = 32.dp),
-                contentAlignment = Alignment.Center,
-            ) {
-                Text(
-                    text = stringResource(R.string.all_setting),
-                    color = Gray4,
-                    style = HearitTypoGraphy.titleLarge,
-                )
-            }
-        },
-        navigationIcon = {
-            IconButton(onClick = onBackClick) {
-                Icon(
-                    painter = painterResource(id = R.drawable.ic_back),
-                    contentDescription = "뒤로가기",
-                    tint = Gray4,
-                )
-            }
-        },
-        colors =
-            TopAppBarDefaults.topAppBarColors(
-                containerColor = HearitBlack,
-            ),
-    )
-}
-
-@Composable
-private fun SettingsContent(
-    modifier: Modifier = Modifier,
+fun SettingContent(
     userInfo: UserInfo,
     onProfileClick: () -> Unit,
     onPrivacyPolicyClick: () -> Unit,
     onTermsClick: () -> Unit,
     onOpenSourceClick: () -> Unit,
+    onLoginClick: () -> Unit,
     onLogoutClick: () -> Unit,
     onWithdrawalClick: () -> Unit,
+    modifier: Modifier = Modifier,
 ) {
     Column(
         modifier = modifier.fillMaxSize(),
@@ -151,7 +71,7 @@ private fun SettingsContent(
         } else {
             SettingItem(
                 text = stringResource(R.string.all_login),
-                onClick = onLogoutClick,
+                onClick = onLoginClick,
             )
         }
 
@@ -166,7 +86,7 @@ private fun SettingsContent(
 }
 
 @Composable
-fun SettingItem(
+private fun SettingItem(
     text: String,
     onClick: () -> Unit,
     modifier: Modifier = Modifier,
@@ -188,29 +108,17 @@ fun SettingItem(
     }
 }
 
-private fun navigateToLicense(context: Context) {
-    OssLicensesMenuActivity.setActivityTitle(context.getString(R.string.oss_license_title))
-    val intent = Intent(context, OssLicensesMenuActivity::class.java)
-    context.startActivity(intent)
-}
-
-private fun openUrl(
-    url: String,
-    context: Context,
-) {
-    context.startActivity(Intent(Intent.ACTION_VIEW, url.toUri()))
-}
-
 @Preview(showBackground = true)
 @Composable
 private fun RegisteredSettingScreenPreview() {
     MaterialTheme {
-        SettingsContent(
+        SettingContent(
             userInfo = UserInfo.default(),
             onProfileClick = {},
             onPrivacyPolicyClick = {},
             onTermsClick = {},
             onOpenSourceClick = {},
+            onLoginClick = {},
             onLogoutClick = {},
             onWithdrawalClick = {},
         )
@@ -221,12 +129,13 @@ private fun RegisteredSettingScreenPreview() {
 @Composable
 private fun NotRegisteredSettingScreenPreview() {
     MaterialTheme {
-        SettingsContent(
+        SettingContent(
             userInfo = UserInfo(id = 0L, nickname = "hEARit", profileImage = null),
             onProfileClick = {},
             onPrivacyPolicyClick = {},
             onTermsClick = {},
             onOpenSourceClick = {},
+            onLoginClick = {},
             onLogoutClick = {},
             onWithdrawalClick = {},
         )
